Add [] call life_fnc_saveGear; to fn_sessionUpdate
For some reason, when player sync'ing is initiated, the cop_gear was not updated.  If this was INTENTIONAL then we will have to reverse this commit.  The only reason I can think why is to allow cops to reconnect or respawn with all their stuff, just as the last time they bought or sold something at a cop shop.  Problem is that intentional item dropping or adding is permanent only if they buy it.  Could be abused to item duplicate.  This behavior can be controlled by a global configuration variable if need be.