Beginnings of civ friendly IEDs.
WEST will have to have a mine detector and be on foot to get a 20meter
detection radius.  Toolkit necessary to defuse (undecided whether it
should be consumed in defusing)
Driving within 10m of the IED will set it off.  Blast will be
non-discriminatory, but only a vehicle with a WEST inside can set it
off.
This is not complete or working yet.